pinglab/pyproject.toml
[project]
name = "pinglab"
version = "0.1.0"
description = "pinglab library"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.10.0",
    "numpy>=2.2.6",
    "pydantic>=2.12.4",
    "pyyaml>=6.0.2",
    "tqdm>=4.67.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

